Explore most likely optimal choice first
We switch the exploration selection engine.
It used to pick systematically the one which has the best lower bound.
However, that was both very arbitrary
(as we pick an arbitrary lower bound
from a rule-of-succession style weighing),
and more importantly for now,
that multi-armed bandit selection method
can lead to many possible choices being unexplorable,
if their lower bound will never be below the lower bound
of the current best choice
(whose lower bound gets higher as precision increases).

The new engine relies on probability of selection
based on an estimation of the probability
that a given choice is optimal,
and on the likelihood that the choice can be improved from exploration.
The optimality probability is a very rough draft,
using a logistic distribution: we pretend that the choice is a Gumbel
whose variance we extract from two samples:
the current mean, and the lower bound
(computed as with the previous engine).
Then we compare it by subtraction from the Gumbel distribution
of the current best choice,
and compute the probability that it is negative.

Despite the gross approximation,
this allows us to reach 3.4242 average guesses in 1K explorations,
or 7 guesses from perfect play.