Estimate probability of self-improvement statistically
Previously, we computed the probability of improvement
as the quotient of improvements among explorations.

However, the probability that we actually care about is the probability
that some future exploration, at any later point,
will yield an improvement (instead of just the next exploration).
Indeed, the probability that we want to select for,
is the likelihood that the explored choice will beat the current best.

We estimate that probability from a (for now heavily approximate)
estimation relying on considering explorations that lead to an
improvement as a Gumbel distribution, and the optimal value as another,
and computing the probability that it will beat it logistically.

It achieves 3.4216 average guesses to win in 1K explorations,
or 1 more guess than optimal.
It achieves perfect play in 1234 explorations.