Estimate sample variance from observed data
This finally frees us from the last arbitrary, Wordle-specific
code: the lower bound confidence computation relied on an arbitrary
improvement of 0.07 average guesses to win with one more exploration.
That value worked surprisingly well, but it is a local optimum,
and we want to generalize the algorithm to any minimization turn-based
problem.

We also improve the improvement probability estinate by relying on a
Gumbel distribution (which uses the sample variance) to find out whether
the optimal estimate is lower than the best measurement yet.

One negative is that there is a substantial decrease in convergence
speed, and we no longer reach optimality in a reasonable time.
We expect that it is a way to a superior convergence once we improve the
optimality estimate further, along with the probabilities.