Improve optimal value estimate from two-sided derivates
Previously, we attempted to infer optimal estimates from the local
derivative at the current number of visits in the explored choice.
That lead to very highly incorrect values.

In order to ground the estimate, we rely on both the measured derivative
at the start of exploration, and that at the current number of visits,
from which we can infer higher-quality parameters for the exponential
plateauing of exploration improvements across all values.

We take this opportunity to also use those parameters on other choices
which have only seen a single visit.