Compute optimal probability from besting all other choices
We assume that besting one choice is independent of besting another.
We compute the probability of besting a choice from their respective
variance (computed from the average sample prediction variance) and
assuming each choice’s optimal estimate takes a Gumbel distribution,
such that their difference is logistic.
The probability of besting another choice is the integral of the
logistic for a positive difference between their optimal values.

This yields the optimal policy in 1h10, 918 steps.