Compute optimality probabilities from all choices and precisely
Prior computations were inexact because they relied on updating the
denominator of the probability quotient: the sum of probabilities that
each individual choice is superior to the others in some sense.
(Through imprecision, that sum did not equal one,
making this quotient necessary.)
However, the update of the denominator only substituted the old
optimality probability for the explored choice, against the new one.
Yet the optimality probability of the other choices were affected by
this change: indeed, they all depend on the optimal value estimate
and its associated variance, which typically change after each
exploration.

Thus we recompute them all each time.
This ends up not costing that much more CPU, likely because the
RAM allocation costs dwarf other latencies.

Another improvement we include is in the algorithm for optimality
probability. Previously, we were computing the probability that the
current choice would beat the current best choice. However, that results
in all choices’ probabilities being very similar; all between 30% - 70%.
The division by the sum resulted in exploration probabilities nearing
1%, which effectively acted like a round robin over all choices
regardless of realistic probability of optimality.

We now assume that all choices have independent optimality
probabilities. We compute the optimality probability as the probability
that the choice beats *all* other choices, instead of just the best one.
The resulting selection system feels extremely coherent, contrary to
what we had previously (which was either too shy to pick better choices,
or quickly stopped visiting anything but the estimated best).

That results in a decrease in performance: we reach perfect play after
2h, instead of 1h10 previously (with a selection that stuck too early
with what happened to be the best choice). We see it as a net win.

As a quick note on other solutions we could have implemented:
a common industry practice relies on softmax to force the categorical
set of choices into a near-one-hot vector. However, that takes no
account to individual, uneven variances of the choices, and generally is
a bit of a hack. The cost of multiplying through all choices is not
currently an issue.