Support all guesses, change optimal probability
We used to cap the number of guesses we explored to only include
the 100 choices with the best estimates.
It was necessary to ensure the program would not be too slow,
and more importantly, would not eat too much RAM
(which would cause OOM kills otherwise).
However, a choice that was not in the top 100 could potentially end up
being optimal, and thus, the algorithm does not technically converge
towards the optimal choice in the general case.

We now remove this cap.
We avoid the RAM cost by having a cache of the N best optimal estimates.
We add a guess to the cache with a frequency matching the probability
that the optimal choice is in the cache,
which we compute when there is no choice in the cache
that we should select based on their probabilistic frequencies.

We also improve the computation of the probability
that a choice is optimal.
Previously, it was computed as the product of the probabilities
that the choice beats each other choice.
However, this is an inaccurate computation,
since the product is only accurate if each event is independent;
yet the probability that the best choice beats the worst
is highly dependent on the probability that it beats the second best;
all those probabilities have transitive properties.

The runtime is now expectedly unusably slower.
The main reason
(besides the obvious loops through the whole guess space)
is that the measurement variance now comes
from the variance of the estimates of all guesses.
Ideally, each variance should be computed separately for each source.
We plan on fixing that next.