Extract optimal estimation from subtrees
Before, we relied exclusively on raw, “real” measurements of the average
number of guesses for a given choice using the current best policy,
to estimate the optimal number of guesses for a choice.

We now avoid digging through the whole subtree, only looking one level
down and gathering information from the subnode estimates.

The performance of each search is very slightly improved, although the
number of steps has gone up. We currently do not reach optimal play in a
reasonable timeframe.