Extract logic for convergent measurements
We previously tied the logic for the computation of the estimation for
where the set of measurements for the average number of choices
necessary to solve a given word, would eventually converge to.

We now extract that logic to an independent, self-contained and easier
to study object.

For now, that object no longer learns from the measurements of other
sibling choices. Thus the performance is decreased. We will improve it
in the future in a more generic manner. In particular, we hope that this
more generic construct will help gather more information on the variance
estimation.