Use subtree information in improvement probability
When available, we extract improvement probabilities from subnodes
recursively. The hope is that the improvement probability will be more
precise, especially in the later stages, when most of the leaves have
been explored, and an improvement at that level is unlikely, making an
improvement in parent nodes unlikely despite a recent improvement (which
resets the improvement probability to 1).