Add constraints to breadcrumbs
We previously were logging only the sequence of guesses, which
we printed for debugging purposes whenever an improvement was found.

However, it was confusing to read:
we could see a switch to "salet boeuf" lead to an improvement to 2 guesses,
followed by a switch to "salet routh" which "regressed" to 4 guesses,
as the improvement was for different constraint responses.

Now we can see that one improvement is for "salet x.x.. boeuf",
while the other is for "salet ....x routh".