Factorize computed differentials for all tree choices
As mentioned in d91cc, we had lost the ability to learn from the
measurement slope of other choices as they gather more information.
That is unfortunate, as all choices typically have the same initial bias
which can be compensated for.

We now compute these differentials in a dedicated object that is shared
and averaged among all choices in the tree.

The initial result is OK, but not great. It is not as good as 966fa,
which is not fundamentally surprising, since it gathered feedback from
real measurements of completed games, which can only improve and never
worsen, which forces a one-sided bias to exist.
Meanwhile we are only gathering information from children choices.
Still, we must investigate why the differentials struggle to extrapolate
the biases across choices.