Register average diffs between measurements for successive visits
The main issue with our asymptotical extrapolation,
in the context of a measurement refinement from the children nodes,
is that the average of exponential parameters is heavily sensitive to
the outliers, causing a regression curve which can be significantly off.
That causes extrapolated optima to be incomparable between choices that
have had a different number of visits.

Instead, we now maintain the list of differences between successive
measurements. Whilst the actual values of each measurement is
meaningfully distinct between choices as a consequence of their varying
worth, the bias inherent in their estimation is typically correlated at
each new measurement. Thus we can have a very good estimate of the error
between one measurement and its eventual converged value at a later step
by summing the known average differences between measurements.

We use this as asymptotic estimation.

While this gathers a lot more data (and an O(V) amount, with V the
number of measurement visits, instead of the O(1) we had), it is only
gathered at the level of the tree, instead of for each choice, which
remains an acceptable RAM load.