Build exponential regression off bias slopes
In order to avoid the flaw described in b1de9 while still using
exponential regression, we can compute the exponential parameters based
on the slope of the first half of measurements, and the slope of the
second half, which is much less sensitive to extrema as was using only
the slope at 0 and at the end.
The exponential curve would then fit the data points much better.

We use the list of slopes at each measurement to make that computation,
and compute exact curve parameters every time, avoiding the issues we
had with averaged parameters.

The goal is to use this better system for later switching back to an
estimation that uses full-depth measurements (to the end of the game).
Indeed, refinements from children nodes yield a slow convergence,
which is not balanced by a sufficiently faster iteration.

Surprisingly, while it is not intended to yield good results for
refinement from children nodes, the results are better, likely because
of the higher asymptotic variance.