Hinge computation on the measurement structure
Some elements of the optimizing process were still relying on
measurements outside of the externalized estimators we built.
In particular, the computation of the refined estimation from the
children used the subchoice with the highest lower bound, instead of the
one with the best asymptote.

We now reserve the use of the lower bound to pure information reporting;
it does not participate in the optimization process.

We also switched the improvement estimator to rely exclusively on the
children nodes: it used to rely on the last improvement, which relies on
the visit for which the lower bound bumped up.
However, an investigation on improving the improvement probability is
needed, since it is now stuck at 100%, as a result of the leaves' zero
probability of improvement not being accounted since they are not
represented as choices.