Gumbel-based improvement probability default
Instead of computing the improvement probability of visiting a choice,
from the number of consecutive visits without improvements, which was a
bit hacky and caused the probability to jump to 100% every time an
improvement was found, we instead estimate it from two sources:

- First, from a Gumbel-distribution probability of the asymptote
  actually being the measured lower bound (ie. the known current score
  we currently achieve), given the uncertainty of the asymptote.
- Second, once subnodes are expanded, we combine their improvement
  probabilities recursively, weighed by the likelihood that they are
  explored.

One issue we currently have is that the main toplevel choice, crate,
gets a high optimal probability and keeps it throughout.
The cause is likely the combination of two things:

1. Early on, the exponential regression is a poor fit, yielding unfair
   asymptote estimates.
2. Later on, the variance is lower than it should, because it is
   extrapolated as a power regression from the first two values,
   cementing its top position as certain despite its volatility.

For an unknown reason, its improvement probability converges to 57%.
That prevents this change from improving on the current best.