Compute improvement probability from Gumbel CDF
In 93b9b0, we used the Gumbel distribution PDF to initially estimate the
improvement probability of a choice when no subnodes are present.

However, the PDF gives the probability that the asymptote is exactly at
the lower bound, but there are probabilities that it is lower than the
lower bound which are unaccounted for. Of course, we know that the
asymptote cannot be below the lower bound; to avoid losing those
probabilities, we sum them by using the CDF instead.