Sample variance of historical asymptote estimates
This implements option 2 of commit 513b9.

It helps salet reach 3.4251 in 176 steps instead of 1077 with option 3.
The most likely reason for the substantial improvement is that the
latest variance is not stuck at 0 most of the time.

Another improvement, which would have worked for either option,
is to compute a quadratic regression to smoothe the variance with
respect to the visit count.

The reason we switched to option 2 prior to implementing that regression
is twofold:
- The variance estimation is numerically stable.
- The computation effort is also lowered, although it is not felt much.