Set prob_superior_choice to 0 with inferior asymptote
Previously, it returned 1 when the variance was 0, even when the
asymptote of the choice it compared against was worse.
That resulted in leaf nodes considering that better asymptotes that had
no variance (since, being a leaf, it already reached its maximum) were
considered on par with worse choices.
We were able to notice thanks to 952a42a.