Select in proportion to improving over the best reward
We try out a selection method that, instead of selecting in proportion
to the probability of choices of being optimal, selects them based on
the probability that they improve upon the current best cumulative
reward (for what is considered the best choice).

Selecting based on the optimal probability has the effect of locking the
exploration to the best choices at the expense of unknown (but likely
worse) choices, despite some of them actually being better. The addition
of the improvement probability to discourage exploring stagnating
options, has not been successful so far, despite experiments with it.
The improvement probability just does not go down as fast as the optimal
probability gap goes up. Eventually, it would cause the exploration
probability to converge to zero, causing it to explore the uncached
actions. But at the lower leaves, it is not even close to going over the
edge after a week of computation.