Switch action selector to max exploratory reward
The previous selector was proportional to the probability that a given
action surpassed the current maximum play reward.
However, it caused overselection of unimproving choices simply because
even a best choice that does not statistically improve, is expected to
have a true maximum above its current sample explorations size
(estimated for instance as (sample max) × (size + 1) ÷ size - 1).

We experiment here with something closer to the textbook definition of
Thompson sampling: we select the choice whose expected reward is
maximized. However, unlike our earlier computation of the optimal
probability of each choice, we define the exploratory reward as separate
from the play reward: the former is the improvement in the best play
reward after exploring a given action.

Ideal sampling would select an action in proportion to the probability
that its exploratory reward is maximized across all possible means and
variances of each action.  Because that is not computationally
tractable, we instead perform Monte Carlo sampling of the expected
exploratory reward of each action, and pick the best.

For now, it does not seem to yield superior results (although they do
continuously improve), likely because the estimation of reward of
uncached choices is flawed. Indeed, for now, we simply don’t estimate
it, and merely uncache a choice every time we explore the most recently
uncached choice, which is pragmatically reasonable but not
mathematically justified.