Recompute raw bias and variance for estimates
The variance computation was likely overestimated.
Indeed, it is the variance of the set of estimates
for the current choice.
However, early estimates are more wrong that recent estimates.
While this was known, it was deemed close enough,
especially considering that the estimates converge fairly quickly.

Nevertheless, convergence towards optimality being still slow,
we study whether it would improve from as precise a computation
as we can make it, regardless of how slow that would be.
In particular, the variance of the latest estimates are still very high,
which might make them selected too often
(trusting that they will improve, despite them making no improvements).
The blind estimate (the one without subtree estimates) is also very high,
causing a frequent selection of brand new choices even when stronger
choices still have a lot of opportunity for improvement.

(As you might be able to tell, it is very difficult now to point to a
particular value in the algorithm as the cause of slow convergence.
It is hard to trust any value as right, since some are approximations,
but they all look mostly reasonable. Hence our exploring alternatives.)

This variation is intended to increase our trust in the computed
parameters for the probability distribution of play reward estimates.
It is mostly unsuccessful: the variance computation seems now too
unbalanced between well-explored and seldom-explored choices,
likely because we are pitting choices where the variance is computed
from a bias which includes the distance from the blind estimate to a
very refined and precise estimate (after many explorations),
against choices where the variance involves a bias across few
explorations (thus the estimate assumed to be asymptotic, is likely
wrong).

Nevertheless, this rewrite is a useful step towards more useful
and reliable statistics.