Recompute the best choice exactly
Previously, we computed the best choice
by swapping in the choice that was just explored
if its debiased action value was higher than the previous best choice.
However, changes to tree-level bias information
can also change the debiased action value estimate
of choices that were not explored, such that the “best choice” was not
the one with the highest debiased action value.

We go through all choices and find the one with the best action value.

It matters for convergence, because the best choice of the subtrees
is what is used to compute the tree action value estimate.

It does lead to faster convergence, and the lower bound action value
also converges significantly faster, reaching the debiased estimate
within a thousand steps.