Remove action-value estimate history from choice
To reduce the memory load, we seek to remove the history of tree
estimates that were kept in each choice.
We use a different approach in two places where those were needed:
- We compute the bias in the tree estimate between two visit counts,
  by updating a streaming mean in the tree node.
- We compute the debiased estimate’s variance as the variance
  of the delta between two consecutive debiased estimates,
  across choices. This is an approximation, but it seems OK.

We take this opportunity to improve the log rendering of the tree
as a table, since it now has a few more pieces of information.

We log the output in log/thompson-without-estimate-history.