Fix outdated debiased estimate
We used to incorrectly update the tree node’s bias variance based on the
delta between the old tree estimate with the old bias applied to it,
and the old tree estimate with the new bias applied to it.
Indeed, we believed the update of debiased values in update_tree_bias!()
made the explored choice update its debiased value,
but its tree estimate only got set later in update_action_value!().

That caused the delta to be zero even though there actually was a change
in debiased estimate after a visit. Generally, it caused the variance
estimation to be lower than it should be, reducing the probability of
explorations of a wider range of choices, which caused slow convergence.

We now recompute the new debiased estimate explicitly
and use it in the variance update.