Probabilistic Thompson sampling for exploration
The fair sampling ensures overall visit counts match the optimal
probability, which sometimes cause it to focus on a single choice which
suddenly has a high probability, for a very long time, reducing the
exploration variety, which has a negative impact on convergence time.

We implement two approaches targeted at having an exploration frequency
that matches the optimal probability:
- Explicit frequency matching, where we pick choices
  when their last visit is beyond what the optimal frequency should be.
  The negative consequence is that,
  when a choice has >0.5 optimal probability,
  it is picked half of the time,
  and the second best choice is picked the other half,
  which means other choices are almost never picked
  (the only saving grace being the Monte Carlo probability sampling).
- Probabilistic sampling, where we throw a weighed dice
  and pick the selected choice.
  While it can pick worse choices sometimes,
  it is simple and effective.

We now achieve the optimal strategy in 44 minutes (1334 steps)
on a single thread with no hyperparameters nor game-specific code.