Estimate debiased variance from future deltas
In f4f757, we used to rely on a very rough approximation
for the variance of the debiased action value,
relying only on the delta between two debiased estimates.
The mathematical justification for it is to pretend
that the new debiased estimate approximates the action value,
which is often a poor approximation, as it converges way later.

Here, we perform a more accurate estimation of the variance,
as the sum of the variances of the future deltas
between the debiased action value of two successive visits,
for which a mathematical justification is provided in comments.

- It yields a more accurate result,
- It still does not require storing historical tree estimates,
- It also ensures that the variance of choices with more visits
  are never higher than those with less
  (which used to happen, but is illogical.)
- Qualitatively, it keeps into consideration other close contenders
  for optimal choices longer, and explores more actions,
  which seems more intuitive than the sudden focus on the winner
  that happened before (and which felt luck-based),
  at the expense of a slower convergence.

It reaches optimal play after 4897 steps, in 7h.