Switch to Gaussian Monte-Carlo sampling
Empirically, it finds better strategies faster.

Assuming that the distribution of the action values was a Gumbel,
was a guess anyway, and one which seems to encourage exploring less and
less likely initial guesses at a frequency far above that of those new
guesses becoming contenders for the new optimal guess.