Parallelize search
To avoid biases from appearing, we maintain the selection algorithm;
when the same choice is picked twice, the other thread is blocked
from reading the corresponding search tree node
until the first thread digs into a subtree.

There are slim gains in speed of about a second per iteration.
Given the average of 9s/iter, this is minor.