Set variance to latest nonzero value when it is zero
It is very common for the action with the most visits to have a zero
variance. While the algorithm works mostly fine,
it is obvious that the true variance of the action value is nonzero
(since subsequence visits still find better results).
Worse, in situations where two action have the same value,
only the first gets explored, since their variances are zero.
That has an even bigger impact when the second one is actually optimal,
and more visits (which it never receives) would uncover a higher value.

This is what happens on the action `tarse ..x.x build`
with the averaged-remaining-solutions estimator.
With this change, it goes from reaching an optimal value of 3.5532
to 3.5519, a single guess away from the optimal 3.5516.

(The last move it would need to find, `tarse ....x pilon`,
is sadly not explored because its probability is judged too low.)