Use minimum frequentist visit in action exploration
The code is simpler, inspired by previous commits on UCB,
and the theoretical basis is more solid than picking the first action
that is below the quantized threshold for visits.

The reality is that the frequentist approach
has a continuous, real-numbered visit,
not a discrete, integer one.
What we did before,
jumping out of the loop when we reached the discretized threshold,
had a behaviour that depended on the order of the actions in the list.
Using the smallest frequentist (continuous) visit
no longer depends on it.

However, there remains a flaw in the frequentist sampler
compared to the probabilist one
when one action has a high optimal probability.
Indeed, if the probability of an action is 75%,
it will select that action every other simulation,
as if the probability was 50%.
The probabilist sampler correctly matches the probability.

It may explain why empirically,
the frequentist approach stops behaving as well as the probabilist one
after we reach a certain number of simulations,
for which we start having an action
that is optimal with high confidence.

This approach does not work for the fair sampler,
because the ratio of action visits to state visits
is always zero for a fresh action,
and thus minimal across actions.
It hints at the fact that the fair sampler is biased,
which its poor empirical performance highlights.